+++ src/config/index.ts
@@ -1,5 +1,6 @@
 export * from './app';
 export * from './auth';
+export * from './classifier';
 export * from './code';
 export * from './constant';
 export * from './cors';